Remove pip package (not pipx)
