+++ tests/package/test_pip.py
@@ -1,28 +1,27 @@
-from abc import abstractmethod
 from pathlib import Path
 from typing import Optional
 
 import pytest
 
 from mybox.driver import LocalDriver
-from mybox.package import PipPackage
+from mybox.package import PipxPackage
 from mybox.state import DB
 
 from .base import DOCKER, PackageArgs, PackageTestBase
 
 
-class NoPypiPipPackage(PipPackage):
+class NoPypiPipxPackage(PipxPackage):
     async def _get_pypi_version(self) -> Optional[str]:
         return None
 
 
-async def _test_pip_version(pip_class: type[PipPackage]):
-    package = pip_class(pip="django", driver=LocalDriver(), db=DB(":memory:"))
+async def _test_pip_version(pip_class: type[PipxPackage]):
+    package = pip_class(pipx="django", driver=LocalDriver(), db=DB(":memory:"))
     version = await package.get_remote_version()
     assert version >= "4.1.5"
 
     non_existent = pip_class(
-        pip="xxxxxxxxxxxx", driver=LocalDriver(), db=DB(":memory:")
+        pipx="xxxxxxxxxxxx", driver=LocalDriver(), db=DB(":memory:")
     )
 
     with pytest.raises(Exception, match="Cannot find latest version"):
@@ -31,22 +30,17 @@ async def _test_pip_version(pip_class: type[PipPackage]):
 
 @pytest.mark.trio
 async def test_remote_version():
-    await _test_pip_version(PipPackage)
+    await _test_pip_version(PipxPackage)
 
 
 @pytest.mark.trio
 async def test_remote_version_from_index():
-    await _test_pip_version(NoPypiPipPackage)
+    await _test_pip_version(NoPypiPipxPackage)
 
 
-class DjangoTestBase(PackageTestBase):
-    @property
-    @abstractmethod
-    def installation_method(self) -> str:
-        pass
-
+class TestPipx(PackageTestBase):
     async def constructor_args(self) -> PackageArgs:
-        return {self.installation_method: "django"}
+        return {"pipx": "django"}
 
     async def check_installed_command(self):
         return ["django-admin", "help"]
@@ -59,24 +53,6 @@ class DjangoTestBase(PackageTestBase):
         await super().check_applicable()
         if not DOCKER:
             pytest.skip("Cannot test pip packages without Docker.")
-
-    async def ignored_paths(self) -> set[Path]:
-        return await super().ignored_paths() | {
-            await self.check_driver.home() / ".local" / "bin",
-            await self.check_driver.home() / ".local" / "lib",
-        }
-
-
-class TestPip(DjangoTestBase):
-    @property
-    def installation_method(self) -> str:
-        return "pip"
-
-
-class TestPipx(DjangoTestBase):
-    @property
-    def installation_method(self) -> str:
-        return "pipx"
 
     async def ignored_paths(self) -> set[Path]:
         return await super().ignored_paths() | {
